+++ src/egeo-demo/st-horizontal-tabs-demo/st-horizontal-tabs-demo.component.ts
@@ -25,4 +25,6 @@ export class StHorizontalTabsDemoComponent {
       {id: 'dashboard', text: 'Dashboard'},
       {id: 'nodes', text: 'Nodes'}
    ];
+
+   public active: StHorizontalTab = this.options[1];
 }